requirements.txt
python-telegram-bot==20.4
gspread
google-auth